backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
psycopg2-binary>=2.9
marshmallow==3.20.1
python-dotenv==1.0.0
